vectors/requirements.txt
# RAG (Retrieval-Augmented Generation) Demo
# Vector-based document search and generation

# Core RAG dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5

# Vector database
chromadb>=0.4.0
faiss-cpu>=1.7.4  # Alternative: use faiss-gpu for GPU support

# Text processing
unstructured>=0.11.0
pypdf>=3.17.0
python-docx>=0.8.11

# Embeddings
sentence-transformers>=2.2.2
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
tiktoken>=0.5.0

# Optional: For web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Optional: For advanced chunking
spacy>=3.7.0
